Extract to MSG dispatcher
