+++ common/connection/CMakeLists.txt
@@ -3,16 +3,14 @@ set(HEADERS
     include/connection/Receiver.hpp
     include/connection/Socket.hpp
     include/connection/Sender.hpp
-    include/connection/Server.hpp
-    include/connection/Client.hpp
+    include/connection/MsgDispatcher.hpp
 )
 
 set(SOURCE_FILES
     src/Receiver.cpp
     src/Socket.cpp
     src/Sender.cpp
-    src/Server.cpp
-    src/Client.cpp
+    src/MsgDispatcher.cpp
 )
 
 add_library(connection ${HEADERS} ${SOURCE_FILES})
